+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 ## Added
 - setns.py with setns context manager
 - args_aliases.py with dicts of possible additional arguments for many cases
+- Inject class and cli
 
 ## Fixed
 - error handling with ctypes in cloning.py
+- setns
 
 ## 1.2.4 - 2015-05-02
 - docstrings